fix : saveAccount 함수 내부로직으로 분리
